API/requirements.txt
tensorflow==2.15.0
transformers==4.37.2
fastapi==0.110.3
h5py==3.11.0
pydantic==1.10.8
numpy==1.24.3
uvicorn==0.30.1
nltk==3.8.1
pillow==10.3.0
googletrans==3.1.0a0
Sastrawi==1.0.1
python-dotenv>=1.0.1,<2.0.0
gunicorn>=22.0.0,<23.0.0
python-multipart>=0.0.9
firebase-admin==6.5.0
pandas==2.2.2